feat: part 4 Testing Express servers, user administration d finished

+++ part4/bloglist/backend/controllers/blogs.js
@@ -1,22 +1,53 @@
 const blogsRouter = require('express').Router()
-const { application } = require('express')
 const Blog = require('../models/blog')
-
+const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog
+    .find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
 blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
-  const result = await blog.save()
-  response.status(201).json(result)
+  const body = request.body
+  const user = request.user
+  if(!user) {
+    response.status(401).end()
+  } else {
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes,
+      user: user._id
+    })
+
+    const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+    response.status(201).json(savedBlog)
+  }
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const user = request.user
+
+  if(!user) {
+    response.status(401).end()
+  } else {
+    const blog = await Blog.findById(request.params.id)
+
+    if(blog.user.toString() === user.id.toString()) {
+      await Blog.findByIdAndDelete(request.params.id)
+      user.blogs = user.blogs.filter(savedBlog => 
+        savedBlog.toString() !== request.params.id
+      )
+      await user.save()
+      response.status(204).end()
+    } else {
+      response.status(403).end()
+    }
+  }
 })
